feat(host-plugins): expect host plugins to be provided under the `plugins` property
BREAKING CHANGE: host plugins are now expected to be provided under the `plugins` property of the
options object. host are expected to be provided as full plugins. in addition, a host option can no
longer provide a `projectTypes` list to conditionally list the plugins as options. an alternative
might be provided in the future