feat(ci-services-plugins): expect `ciServices` plugins to be provided under the `plugins` property
BREAKING CHANGE: the list of `ciServices` plugins are now expected to be provided under the
`plugins` property of the options object. the plugins are expected to be provided as full plugins. a
ciService can no longer define whether they are applicable to public or private projects. a
replacement ability to filter might be provided at some point in the future